+++ src/components/page_ui/PageLogo.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const PageLogo: React.FC = () => {
+  return (
+    <div className="fixed top-8 left-10 z-20">
+      <h1 className="font-serif text-6xl text-black drop-shadow-md">
+        Overseer
+      </h1>
+    </div>
+  );
+};
+
+export default PageLogo;
