--- src/data/spheres.ts
@@ -1,13 +1,0 @@
-// [x, y, z, hue, saturation, brightness]
-export type Vector6D = [number, number, number, number, number, number];
-
-// Example list of 6D vectors
-export const vectors: Vector6D[] = [
-  [0, 0.5, 0, 0, 66, 77],
-  [0, 0.5, 0.5, 0, 65, 90],
-  [1, 1, 1, 60, 50, 86],
-  [-1, 0, 1, 120, 45, 69],
-  [0, -1, -1, 180, 70, 49],
-  [-1, 1, 0, 240, 69, 69],
-  [1, -1, -1, 300, 56, 35]
-];
